hopefully final version
added plot residuals to eigen program and also some code that shows the
convergence of the first few eigenenergies

time evo method now makes movies but I have to remember to modify a
parameter when I do that. Cant do that and view plots at the same time.
so many other things honestly.